requirements.txt
boto3==1.35.8
Pillow==10.4.0
numpy==2.1.0